+++ msmexplorer/plots/__init__.py
@@ -3,5 +3,4 @@ from .msm import *
 from .tpt import *
 from .projection import *
 from .cluster import *
-from .timeseries import *
 from .misc import *
